File structure revamp
- Separate functions and classes into separate files to make
  maintenance easier

- Implement test_environment.py and associated variables to allow
  easy of switching between live and test environments
