--- Code/Current code/filter_variables/filter_variables.txt
@@ -1,8 +1,0 @@
-134
-57
-55
-261
-186
-123
-245
-17
