+++ Code/ball_recognition_v1/camera.py
@@ -1,0 +1,67 @@
+"""Kuwin Wyke
+Midwestern State University
+Start: 6 November 2019
+End: work in progress
+
+This module is designed to be used in the ball recognition program in the
+folder named "ball_recognition_v1". It contains a class to connect to the
+camera on either a raspberry pi or windows computer.
+"""
+
+import platform
+import cv2
+
+# Project specific modules
+import variables
+from debug import debug_cam
+# Check if running on rasp pi to import pi cam modules
+if platform.system() != "Windows":
+    import serial
+    from picamera.array import PiRGBArray
+    from picamera import PiCamera
+
+
+# Create class to handle camera object
+class Camera:
+    """ This class is used to create a camera object. With this class
+    both windows and raspberry pi will be able to use the same object because
+    this class will handle the differences between the two operating systems.
+    """
+
+    # Initialize the class. No variables are required because the only
+    # condition is the operating system.
+    def __init__(self):
+        # Check if the operating system is linux, (raspberry pi will return
+        # ("Linux"). The raspberry pi uses a pi cam and therefore needs
+        # the pi cam implementation of the camera.
+        if platform.system() == "Linux":
+            if debug_cam:
+                print("starting camera")
+            self.cam = PiCamera()
+            self.cam.resolution = (variables.set_width, variables.set_height)
+            self.raw_capture = PiRGBArray(self.cam)
+        else:
+            if debug_cam:
+                print("starting camera")
+            self.cap = cv2.VideoCapture(0)
+            self.cap.set(3, variables.set_width)
+            self.cap.set(4, variables.set_height)
+        self.get_frame()
+        # Get the actual width and height that was set.
+        self.height, self.width, self.channels = self.frame.shape
+
+    def get_frame(self):
+        """ This function gets the current image in front of the camera"""
+        if platform.system() == "Linux":
+            self.cam.capture(self.raw_capture, "bgr", True)
+            self.frame = self.raw_capture.array
+            self.raw_capture.truncate(0)
+        else:
+            _, self.frame = self.cap.read()
+
+    # Create the destructor to close the camera connections.
+    def __del__(self):
+        if platform.system() == "Linux":
+            self.cam.close()
+        else:
+            self.cap.release()
