+++ Code/ball_recognition_v1/filter_variables.txt
@@ -1,0 +1,8 @@
+151
+57
+52
+256
+181
+120
+34
+21
